refactor(SetupDataLoader): 계층 권한 초기화 리팩터링
