+++ src/main/java/com/soolsul/soolsulserver/user/auth/listener/SetupDataLoader.java
@@ -1,7 +1,7 @@
 package com.soolsul.soolsulserver.user.auth.listener;
 
 import com.soolsul.soolsulserver.user.auth.domain.RoleHierarchy;
-import com.soolsul.soolsulserver.user.auth.repository.RoleHierarchyRepository;
+import com.soolsul.soolsulserver.user.auth.persistence.RoleHierarchyRepository;
 import com.soolsul.soolsulserver.user.auth.vo.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationListener;
@@ -27,33 +27,23 @@ public class SetupDataLoader implements ApplicationListener<ContextRefreshedEven
         alreadySetup = true;
     }
 
+    @Transactional
+    public void createRoleHierarchyIfNotFound(Role childRole, Role parentRole) {
+        RoleHierarchy roleHierarchy = roleHierarchyRepository.findByChildName(parentRole.getRole())
+                .orElseGet(() -> RoleHierarchy.builder().childName(parentRole.getRole()).build());
+
+        RoleHierarchy parentRoleHierarchy = roleHierarchyRepository.save(roleHierarchy);
+
+        roleHierarchy = roleHierarchyRepository.findByChildName(childRole.getRole())
+                .orElseGet(() -> RoleHierarchy.builder().childName(childRole.getRole()).build());
+
+        RoleHierarchy childRoleHierarchy = roleHierarchyRepository.save(roleHierarchy);
+        childRoleHierarchy.setParentName(parentRoleHierarchy);
+    }
 
     private void setupSecurityResources() {
         createRoleHierarchyIfNotFound(Role.USER, Role.ADMIN);
         createRoleHierarchyIfNotFound(Role.ANONYMOUS, Role.USER);
-    }
-
-
-    @Transactional
-    public void createRoleHierarchyIfNotFound(Role childRole, Role parentRole) {
-        RoleHierarchy roleHierarchy = roleHierarchyRepository.findByChildName(parentRole.getRole());
-
-        if (roleHierarchy == null) {
-            roleHierarchy = RoleHierarchy.builder()
-                    .childName(parentRole.getRole())
-                    .build();
-        }
-        RoleHierarchy parentRoleHierarchy = roleHierarchyRepository.save(roleHierarchy);
-
-        roleHierarchy = roleHierarchyRepository.findByChildName(childRole.getRole());
-        if (roleHierarchy == null) {
-            roleHierarchy = RoleHierarchy.builder()
-                    .childName(childRole.getRole())
-                    .build();
-        }
-
-        RoleHierarchy childRoleHierarchy = roleHierarchyRepository.save(roleHierarchy);
-        childRoleHierarchy.setParentName(parentRoleHierarchy);
     }
 }
