+++ src/main/java/com/soolsul/soolsulserver/user/auth/persistence/RoleHierarchyRepository.java
@@ -1,9 +1,11 @@
-package com.soolsul.soolsulserver.user.auth.repository;
+package com.soolsul.soolsulserver.user.auth.persistence;
 
 import com.soolsul.soolsulserver.user.auth.domain.RoleHierarchy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface RoleHierarchyRepository extends JpaRepository<RoleHierarchy, String> {
 
-    RoleHierarchy findByChildName(String roleName);
+    Optional<RoleHierarchy> findByChildName(String roleName);
 }
